Improving organize_from_tractoflow and _from_recobundles

+++ please_copy_and_adapt/organize_from_tractoflow.sh
@@ -1,28 +1,35 @@
-################################################################################
-# This script will create symlinks in dwi_ml_ready, pointing to your data      #
-# from tractoflow for each subject:                                            #
-#    - Resample/dwi_resample                                                   #
-#         Last dwi output from tractoflow.                                     #
-#         Will be copied in dwi/dwi_tractoflow.                                #
-#    - Eddy/bval_eddy / bvec_eddy                                              #
-#         Will be copied in dwi/bval_tractoflow.                               #
-#    - DTI_metrics/fa.                                                         #
-#         Will be copied to dwi/fa.                                            #
-#    - Register_T1/t1_warp                                                     #
-#        Last T1 from tractoflow.                                              #
-#        Will be copied to anat/t1.                                            #
-#    - Segment_Tissues/map_wm.                                                 #
-#        Will be copied to anat/wm_map                                         #
-#    - Segment_Tissues/mask_wm.                                                #
-#        Will be copied to masks/wm                                            #
-#                                                                              #
-# If you need something else for your model, you can modify this script.       #
-#                                                                              #
-# See our doc for more information                                             #
+
+###############################################################################
+# Your tree should look like:                                                 #
+# derivatives                                                                 #
+#    ├── original (ex, tractoflow input)                                      #
+#    └── preprocessed (ex, tractoflow output + recobundle output)             #
+#    └── dwi_ml_ready: will be created now.                                   #
+#                                                                             #
+# This script will create symlinks in dwi_ml_ready, pointing to your data     #
+# from tractoflow for each subject:                                           #
+#       └── dwi:                                                              #
+#           └── dwi_tractoflow: Resample/dwi_resample                         #
+#               (It is the last dwi output from tractoflow.)                  #
+#           └── bval_tractoflow: Eddy/bval_eddy                               #
+#           └── bvec_tractoflow: Eddy/bvec_eddy                               #
+#           └── fa: DTI_metrics/fa                                            #
+#       └── anat :                                                            #
+#           └── t1_tractoflow: Register_T1/t1_warp                            #
+#               (It is the last T1 output from tractoflow.)                   #
+#           └── wm_map: Segment_Tissues/map_wm                                #
+#       └── masks :                                                           #
+#           └── wm: egment_Tissues/mask_wm                                    #
+#       └── bundles  :                                                        #
+#           └── tractoflow_wholebrain: Tracking/tracking                      #
+#                                                                             #
+# If you need something else for your model, you can modify this script.      #
+#                                                                             #
+# See our doc for more information                                            #
 # (https://dwi-ml.readthedocs.io/en/latest/data_organization.html#ref-organization).
-# We suppose that you have a "preprocessed" folder that contains RecobundlesX  #
-# results folder for each subject.                                             #
-################################################################################
+# We suppose that you have a "preprocessed" folder that contains RecobundlesX #
+# results folder for each subject.                                            #
+###############################################################################
 
 # =====================================#
 #  VARIABLES TO BE DEFINED BY THE USER #
@@ -31,8 +38,8 @@
 #     folder. Will eventually contain dwi_ml_ready/.
 # - subjects = The list of ALL subjects. You may choose later which ones will be
 #     in your training set/validation set/testing set. One subject per line.
-database_folder=YOUR WORKING FOLDER
-subject_list=SUBJECTS.txt
+database_folder=$1   # YOUR WORKING FOLDER
+subject_list=$2      # ex, SUBJECTS.txt
 
 # =====================================#
 #            MAIN SCRIPT               #
@@ -65,7 +72,9 @@ fi
 
 # Reorganizing all subjects
 echo "Checks passed. Now reorganizing subjects"
-while IFS= read -r subjid; do
+subjects=$(<$subject_list)
+for subjid in $subjects
+do
   echo "Reorganizing subject $subjid"
   subj_preprocessed_folder=$database_folder/preprocessed/$subjid
   subj_folder=$dwi_ml_ready_folder/$subjid
@@ -75,25 +84,30 @@ while IFS= read -r subjid; do
   mkdir $subj_folder/masks
   mkdir $subj_folder/bundles
 
-  echo "creating symlinks"
   # dwi:
+  if [ ! -f $subj_preprocessed_folder/Resample_DWI/${subjid}__dwi_resampled.nii.gz ]; then echo "Subject's DWI not found"; exit 1; fi
   ln -s $subj_preprocessed_folder/Resample_DWI/${subjid}__dwi_resampled.nii.gz $subj_folder/dwi/dwi_tractoflow.nii.gz
+  if [ ! -f $subj_preprocessed_folder/Eddy/${subjid}__bval_eddy ]; then echo "Subject's bval not found"; exit 1; fi
   ln -s $subj_preprocessed_folder/Eddy/${subjid}__bval_eddy $subj_folder/dwi/bval_tractoflow
+  if [ ! -f $subj_preprocessed_folder/Eddy/${subjid}__dwi_eddy_corrected.bvec ]; then echo "Subject's bvec not found"; exit 1; fi
   ln -s $subj_preprocessed_folder/Eddy/${subjid}__dwi_eddy_corrected.bvec $subj_folder/dwi/bvec_tractoflow
+  if [ ! -f $subj_preprocessed_folder/DTI_Metrics/${subjid}__fa.nii.gz ]; then echo "Subject's FA not found"; exit 1; fi
   ln -s $subj_preprocessed_folder/DTI_Metrics/${subjid}__fa.nii.gz $subj_folder/dwi/fa.nii.gz
 
   # anat:
-  ln -s $subj_preprocessed_folder/Register_T1/${subjid}__t1_warped.nii.gz $subj_folder/anat/t1.nii.gz
+  if [ ! -f $subj_preprocessed_folder/Register_T1/${subjid}__t1_warped.nii.gz ]; then echo "Subject's T1 not found"; exit 1; fi
+  ln -s $subj_preprocessed_folder/Register_T1/${subjid}__t1_warped.nii.gz $subj_folder/anat/t1_tractoflow.nii.gz
+  if [ ! -f $subj_preprocessed_folder/Segment_Tissues/${subjid}__map_wm.nii.gz ]; then echo "Subject's WM map not found"; exit 1; fi
   ln -s $subj_preprocessed_folder/Segment_Tissues/${subjid}__map_wm.nii.gz $subj_folder/anat/wm_map.nii.gz
 
   # masks:
+  if [ ! -f $subj_preprocessed_folder/Segment_Tissues/${subjid}__mask_wm.nii.gz ]; then echo "Subject's WM mask not found"; exit 1; fi
   ln -s $subj_preprocessed_folder/Segment_Tissues/${subjid}__mask_wm.nii.gz $subj_folder/masks/wm.nii.gz
 
   # bundles:
-  ln -s $subj_preprocessed_folder/Tracking/${subjid}__tracking.trk $subj_folder/bundles/tractoflow_wholebrain.trk
+  if [ ! -f $subj_preprocessed_folder/*Tracking/${subjid}__*tracking*.trk ]; then echo "Subject's tractogram not found"; exit 1; fi
+  ln -s $subj_preprocessed_folder/*Tracking/${subjid}__*tracking*.trk $subj_folder/bundles/tractoflow_wholebrain.trk
 
-done < $subject_list
+done
 
-echo "We have organized tractoflow results into dwi_ml (dwi, anat, masks)".
-echo "We do not raise warnings if one file is not found. Please check that all data was indeed found."
-echo "Ex: 'for subj in dwi_ml_ready/*; do ls \$subj/dwi/*bvec*; done'"
+echo "We have organized tractoflow results into dwi_ml (dwi, anat, masks)"
